inspektor: Introduce license functionality
Make it possible to verify the lack/add license
headers to python files that don't have it. The
current implementation isn't the smartest ever,
which means if the header in the file isn't exactly
like our snippets, it'll try to add a 2nd header,
so the heuristics of figuring out wether the file
already has a license header has to be refined.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@redhat.com>