Ignore filenames which are not the proper source code.
By default, ignore filenames that ends with ~, #, .swp, .pyc, .pyo, .o
since they are not the proper source code.

If the current directory has the .gitignore file or
$GIT_DIR/.config/gitignore or even $HOME/.config/git/ignore,
then this file (the first available) is read as a glob pattern and
extend the ignore list.

Note that, when performing the walking on files, the ignore list
is used to find directories to ignore (say build/).

Signed-off-by: Rudá Moura <rmoura@redhat.com>