Module import strategy changes: avoid unnecessary imports
Right now, the outer level modules will import the inner level modules
by default.  For instance, `inspektor` will import `inspektor.cli`,
which will in turn import `inspektor.cli.app`.

It looks like the reason for doing so is that once `inspketor` is
imported, everything else will be already loaded.  But, I don't
see a reason for that.  Additionally, I changed the imports to
be relative.

Finally, the real motivation for this change is that it makes
inspketor Python 3 compatible.

Signed-off-by: Cleber Rosa <crosa@redhat.com>