inspektor: Introduce PathRegistry and change methods to properties
We could use more caching of previous computations made
when running multiple checks in sequence. Let's introduce
a PathRegistry that remembers the path objects whenever
we check them with PathChecker. This will allow us to
be able to reuse heuristics passing from Linter to StyleChecker,
for example.

Also, use properties when possible, as that also helps us
to cache computations and avoid us repeating the same
checks over and over again.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>