inspektor: Use pylint help text to find appropriate flags
Let's use a heuristic that tends to be more reliable instead
of looking for python versions or pylint version strings:
Check directly the --help output. Since in most py apps
the output is auto generated from real command line options,
this should be relatively safe.

This solves a problem with moving target options such as
--score and others.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>