RFC: quick and dirty parallel lint
The lint check is by far the lenghtiest static code check on my
workflow.  Let's make it run in parallel and save some precious time.

This is a quick and dirty version, just to prove the point that we
can and should do it.  I reckon this could and should be done at
a different level, and also available for other checks (such as
style, indent, license, etc).

Signed-off-by: Cleber Rosa <crosa@redhat.com>