lint: Allow parallel execution
Pylint supports parallel execution, let's allow it in inspekt as well.
One draw-back of this is that when parallel pylint is enabled, it does
not reports passed tests as usual, therefor there is no simple way of
adding our "Lint: %(path)s %(status)" debug messages. It is not
impossible, but not feasible, therefor this version prints the list of
results after the execution. Also instead of filename it prints module
name (for the same reason).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>