commands: Remove the "--verbose" argument
The "--verbose" argument is already consumed by cliff's App() and is
used to set logging levels. It is not passed to sub-commands and
therefor the removed code in lint did not work.

Using "--verbose" still changes the logging levels, so all the
"checker.log_status()" messages are only printed when "--verbose" is
enabled.

This patch removes the original "verbose" mode for "pylint" because it
does not actually work every since we moved to cliff. Since no one
complained about it I think it's better to remove it and keep the
current behavior which is to run normal pylint and log the
"checker.log_status()" messages in "verbose" mode.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>